104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <inttypes.h>
 /**
  * main - print first 50 fibonacci
  *
@@ -6,17 +7,17 @@
  */
 int main(void)
 {
-	float prev = 1, cur = 2, next = 0;
+	unsigned __int128 prev = 1, cur = 2, next = 0;
 	int i;
 
 	for (i = 0; i < 97; i++)
 	{
-		printf("%.0f, ", prev);
+		printf("%llu, ",(unsigned long long) prev);
 		next  = cur + prev;
 		prev = cur;
 		cur = next;
 		if (i == 96)
-			printf("%.0f\n", prev);
+			printf("%llu\n",(unsigned long long)prev);
 	}
 	return (0);
 }
